// src/shared/constants/RoutePathKeys.enum.ts
export enum RoutePathKeys {
    // Public routes
    homepage = '/',
    aboutUs = '/despre-noi',
    medicalServices = '/servicii-medicale',
    occupationalMedicine = '/servicii-medicale/medicina-muncii',
    generalMedicine = '/servicii-medicale/medicina-generala',
    alternativeMedicine = '/servicii-medicale/medicina-alternativa',
    labTests = '/servicii-medicale/analize-de-laborator',
    preventionPrograms = '/programe-preventie',
    diets = '/programe-preventie/regimuri-alimentare',
    weightControl = '/programe-preventie/control-al-greutatii',
    breastCancer = '/programe-preventie/preventie-cancer-san',
    firstAid = '/programe-preventie/primul-ajutor',
    diabetes = '/programe-preventie/diabet-zaharat',
    promotions = '/promotii',
    contact = '/contact',
    appointment = '/appointment',
    login = '/login',

    // Protected routes
    dashboard = '/dashboard',
    upload = '/dashboard/upload',
}